# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ExamenTAP API",
        "description": "Documentación de la API REST del backend ExamenTAP.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Lista productos con paginación, búsqueda y ordenamiento",
                "description": "Devuelve un listado paginado de productos. Permite buscar por código o nombre. Disponible para cualquier usuario autenticado.",
                "operationId": "productosIndex",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_field",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "created_at",
                            "enum": [
                                "id",
                                "name",
                                "precio",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductoListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crea un nuevo producto",
                "description": "Crea un producto y registra el alta en la bitácora de auditoría. Requiere que el usuario autenticado tenga el permiso de crear en la sección productos.",
                "operationId": "productosStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductoResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos para crear productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/productos/{producto}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtiene el detalle de un producto",
                "description": "Devuelve la información completa de un producto. Disponible para cualquier usuario autenticado.",
                "operationId": "productosShow",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductoResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualiza un producto existente",
                "description": "Actualiza un producto capturando la información anterior antes de aplicar los cambios y registra la edición en la bitácora de auditoría. Requiere que el usuario autenticado tenga el permiso de editar en la sección productos.",
                "operationId": "productosUpdate",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductoResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos para editar productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Elimina un producto",
                "description": "Elimina físicamente el producto de Firebase Firestore. Requiere que el usuario autenticado tenga el permiso de eliminar en la sección productos.",
                "operationId": "productosDestroy",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos para eliminar productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/productos/export/pdf": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Exporta el listado de productos a PDF",
                "description": "Genera un PDF con título, fecha de generación y una tabla con Código, Nombre, Precio y Fecha de creación.",
                "operationId": "productosExportPdf",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Archivo PDF",
                        "content": {
                            "application/pdf": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/productos/export/excel": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Exporta el listado de productos a Excel",
                "description": "Genera un archivo XLSX con las columnas Código, Nombre, Precio y Fecha de creación.",
                "operationId": "productosExportExcel",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Archivo Excel",
                        "content": {
                            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiSuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Operación exitosa."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ApiErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Las credenciales proporcionadas son incorrectas."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ApiValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation error"
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "El correo electrónico es obligatorio."
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
            "ProductoListResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Productos obtenidos correctamente."
                    },
                    "data": {
                        "properties": {
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            },
                            "pagination": {
                                "$ref": "#/components/schemas/Pagination"
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ProductoResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Producto creado correctamente."
                    },
                    "data": {
                        "$ref": "#/components/schemas/Producto"
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "properties": {
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "example": 42
                    },
                    "total_pages": {
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
...
            "StoreProductoRequest": {
                "required": [
                    "name",
                    "precio"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Teclado mecánico"
                    },
                    "precio": {
                        "type": "number",
                        "format": "float",
                        "example": 599.99
                    }
                },
                "type": "object"
            },
            "UpdateProductoRequest": {
                "required": [
                    "name",
                    "precio"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Teclado mecánico"
                    },
                    "precio": {
                        "type": "number",
                        "format": "float",
                        "example": 649.99
                    }
                },
                "type": "object"
            },
...
            "Producto": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "a1b2c3d4e5f6"
                    },
                    "name": {
                        "type": "string",
                        "example": "Teclado mecánico"
                    },
                    "precio": {
                        "type": "number",
                        "format": "float",
                        "example": 599.99
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-08-05T13:00:00+00:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}